+++ src/AWT.js
@@ -30,8 +30,7 @@
  */
 
 import $ from 'jquery';
-import { settings, findParentsWithChild, getBoolean, getAttr, setAttr, checkColor, colorHasTransparency, fx } from './Utils.js';
-import WebFont from 'webfontloader';
+import { settings, findParentsWithChild, getBoolean, getAttr, setAttr, checkColor, colorHasTransparency, appendStylesheetAtHead, fx, log } from './Utils.js';
 
 /**
  * Font contains properties and provides methods to manage fonts
@@ -71,9 +70,8 @@ export class Font {
     // Load own fonts and remove it from the substitution table
     if (options && options.ownFonts) {
       options.ownFonts.forEach(name => {
-        // Check WebFont as a workaround to avoid problems with a different version of `webfontloader` in agora.xtec.cat
-        if (Font.ALREADY_LOADED_FONTS.indexOf(name) < 0 && WebFont && WebFont.load) {
-          WebFont.load({ custom: { families: [name] } });
+        if (Font.ALREADY_LOADED_FONTS.indexOf(name) < 0) {
+          log('debug', `Loading custom web font: ${name}`);
           Font.ALREADY_LOADED_FONTS.push(name);
           delete substitutions[name.trim().toLowerCase()];
         }
@@ -82,7 +80,6 @@ export class Font {
 
     // Add custom font substitutions
     if (options && options.fontSubstitutions)
-      //substitutions = Object.assign({}, substitutions, options.fontSubstitutions)
       substitutions = $.extend(Object.create(substitutions), options.fontSubstitutions);
 
     if ($tree.jquery)
@@ -119,10 +116,10 @@ export class Font {
    * Try to load a specific font from [http://www.google.com/fonts]
    * @param {string} name - The font family name
    */
-  // Check WebFont as a workaround to avoid problems with a different version of `webfontloader` in agora.xtec.cat
   static loadGoogleFont(name) {
-    if (name && !Font.ALREADY_LOADED_FONTS.includes(name) && WebFont && WebFont.load) {
-      WebFont.load({ google: { families: [name] } });
+    if (name && !Font.ALREADY_LOADED_FONTS.includes(name)) {
+      log('debug', `Loading Google web font: ${name}`);
+      appendStylesheetAtHead(`https://fonts.googleapis.com/css?family=${name}`);
       Font.ALREADY_LOADED_FONTS.push(name);
     }
   }
@@ -1003,7 +1000,7 @@ export class Shape {
   static buildShape(data) {
     const shapeType = (data.type === 'rect' && Rectangle) || (data.type === 'ellipse' && Ellipse) || (data.type === 'path' && Path) || null;
     if (!shapeType) {
-      console.log('unknown shape:', data);
+      console.error('unknown shape:', data);
     } else
       return (new shapeType()).setAttributes(data);
   }