+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### v2.3.0 (Not yet published)
 #### Improvements
 - Use [Vite](https://vite.dev/) instead of [Webpack](https://webpack.js.org/) as default bundler.
+- Removed dependency on package [webfontloader](https://github.com/typekit/webfontloader).
 
 #### Bug fixes
 - Corrected MarkDown typo in TRANSLATIONS.md